Removed an uneeded manager
